# Optimize generated subtargets' interaction with `SpecifiedSourceFiles`