Benchmark runs `./pants --no-pantsd lint 'src/python/pants/util/*.py'` 10 times without pantsd and with `Process` already cached.

Before:
```
            Mean        Std.Dev.    Min         Median      Max
real        2.697       0.034       2.652       2.706       2.752
user        2.354       0.025       2.317       2.361       2.392
sys         0.432       0.011       0.420       0.428       0.452
```

After:
```
            Mean        Std.Dev.    Min         Median      Max
real        2.605       0.018       2.588       2.599       2.648
user        2.259       0.015       2.243       2.253       2.296
sys         0.403       0.005       0.395       0.405       0.408
```

This means that we almost never use `SpecifiedSourceFiles`'s logic. Almost always, `SpecifiedSourceFiles` is the same thing as `AllSourceFiles`. The only time that the two diverge is when file arguments cannot use generated subtargets because there are multiple owners of the file.

[ci skip-rust-tests]